refactor: custombox component added to aoierror class
